git-svn-id: file:///home/nbingham/backup/private-repos/trunk/Software/Work/chp@101 404bc64d-9d1b-498e-bf1b-9a6e5dac48b1

+++ src/space.cpp
@@ -149,7 +149,7 @@ space space::operator[](int i)
 
 	list<state>::iterator j;
 	for (j = states.begin(); j != states.end(); j++)
-		result.states.push_back((*j)[i]);
+		result.states.push_back((*j)[(size_t)i]);
 
 	return result;
 }
